backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
pyttsx3
python-dotenv
requests
